hiperf告警修改

+++ src/perf_events.cpp
@@ -1450,7 +1450,8 @@ void PerfEvents::RecordLoop()
             ReadRecordsFromMmaps();
         }
 
-        if ((uint64_t)std::chrono::duration_cast<milliseconds>(thisTime - startTime).count() > count * THOUSANDS) {
+        if ((uint64_t)std::chrono::duration_cast<milliseconds>(thisTime - startTime).count() > 
+        (uint64_t)(count * THOUSANDS)) {
             if (HaveTargetsExit(startTime)) {
                 break;
             }
